Better mcdoc literal number parsing
