+++ packages/mcdoc/src/binder/index.ts
@@ -84,12 +84,11 @@ import type {
 	DynamicIndex,
 	EnumTypeField,
 	Index,
-	LiteralNumberCaseInsensitiveSuffix,
-	LiteralNumberSuffix,
 	LiteralValue,
 	McdocType,
 	NumericRange,
 	NumericTypeKind,
+	NumericTypeKinds,
 	ParallelIndices,
 	PrimitiveArrayValueKind,
 	StaticIndex,
@@ -97,6 +96,7 @@ import type {
 	StructTypePairField,
 	StructTypeSpreadField,
 } from '../type/index.js'
+import type { LiteralNumberCaseInsensitiveSuffix } from '../parser/index.js'
 
 interface McdocBinderContext extends BinderContext, AdditionalContext {}
 
@@ -1215,9 +1215,8 @@ function convertLiteralValue(
 	} else if (TypedNumberNode.is(node)) {
 		const { suffix, value } = TypedNumberNode.destruct(node)
 		return {
-			kind: 'number',
+			kind: convertLiteralNumberSuffix(suffix, ctx),
 			value: value.value,
-			suffix: convertLiteralNumberSuffix(suffix, ctx),
 		}
 	} else {
 		return {
@@ -1230,11 +1229,16 @@ function convertLiteralValue(
 function convertLiteralNumberSuffix(
 	node: LiteralNode | undefined,
 	ctx: McdocBinderContext,
-): LiteralNumberSuffix | undefined {
+): (typeof NumericTypeKinds)[number] {
 	const suffix = node?.value as LiteralNumberCaseInsensitiveSuffix | undefined
-	return suffix?.toLowerCase() as
-		| Lowercase<Exclude<typeof suffix, undefined>>
-		| undefined
+	switch (suffix?.toLowerCase()) {
+		case 'b': return 'byte';
+		case 's': return 'short';
+		case 'l': return 'long';
+		case 'f': return 'float';
+		case 'd': return 'double';
+		default: return 'int';
+	}
 }
 
 function convertNumericType(
